rp2040: correctly size variables to reduce RAM usage

+++ src/portable/raspberrypi/rp2040/rp2040_usb.h
@@ -67,10 +67,10 @@ struct hw_endpoint
 
     // Current transfer information
     bool active;
-    uint total_len;
-    uint len;
+    uint16_t total_len;
+    uint16_t len;
     // Amount of data with the hardware
-    uint transfer_size;
+    uint16_t transfer_size;
     // Only needed for host mode
     bool last_buf;
     // HOST BUG. Host will incorrect write status to top half of buffer
@@ -80,7 +80,7 @@ struct hw_endpoint
     uint8_t *user_buf;
 
     // Data needed from EP descriptor
-    uint wMaxPacketSize;
+    uint16_t wMaxPacketSize;
     // Interrupt, bulk, etc
     uint8_t transfer_type;
